requirements.txt
flake8==2.5.4
Django==1.9.5
uWSGI==2.0.12
GitPython==1.0.2
djangorestframework==3.3.3
psycopg2==2.6.1
django-extensions
markdown
ipython
ipdb
celery
django-celery
django-cacheback[celery]
